implement waitingroom, many people can join
The splashscreen multiplayer button now goes to the inputname screen which can later become a popup like justing wanted to do with the leavescreen. the inputname screen will when put in something sent a message to the server which checks if the name has already been taken (I used to lowercase() aswell). if the name is not taken the player will be added to the waitingroom game. and a message will be sent back to the client initilizing his game id, score and updating his scene. All other players in a waiting room will get another message updating the player list 

- add inputname screenctrl
- link inputname screen
- update css inputname + waitingroom(text_font.css)
- add listview
- add serverMessage EXTRA_PLAYER for updating list
- add clientMessage START_GAME in the switch case this doesn't have a break this is intentional and can be changed. 
- start game will change waitingRoom variable
- add a way to sent to all players now done by changing a boolean